+++ server/camcops_server/tasks/apeqpt.py
@@ -48,7 +48,7 @@ from camcops_server.cc_modules.cc_task import (
 from camcops_server.cc_modules.cc_trackerhelpers import (
     TrackerInfo,
 )
-from sqlalchemy.sql.sqltypes import Integer, DateTime, UnicodeText
+from sqlalchemy.sql.sqltypes import Integer, UnicodeText
 
 
 # =============================================================================
@@ -61,22 +61,42 @@ class Apeqpt(TaskHasPatientMixin, Task):
     """
     __tablename__ = "apeqpt"
     shortname = "APEQPT"
-    longname = "Assessment Patient Experience Questionnaire: For Psychological Therapies"
+    longname = ("Assessment Patient Experience Questionnaire: "
+                "For Psychological Therapies")
     provides_trackers = True
 
-    q_datetime = CamcopsColumn("q_datetime", PendulumDateTimeAsIsoTextColType, comment="Session number")
+    q_datetime = CamcopsColumn(
+        "q_datetime", PendulumDateTimeAsIsoTextColType,
+        comment="Session date/time")
 
     N_CHOICE_QUESTIONS = 3
-    q1_choice = CamcopsColumn("q1_choice", Integer, comment="Enough information was provided", permitted_value_checker=ZERO_TO_ONE_CHECKER)
-    q2_choice = CamcopsColumn("q2_choice", Integer, comment="Treatment preference", permitted_value_checker=ZERO_TO_ONE_CHECKER)
-    q3_choice = CamcopsColumn("q3_choice", Integer, comment="Preference offered", permitted_value_checker=ZERO_TO_TWO_CHECKER)
+    q1_choice = CamcopsColumn(
+        "q1_choice", Integer,
+        comment="Enough information was provided (0 no, 1 yes)",
+        permitted_value_checker=ZERO_TO_ONE_CHECKER)
+    q2_choice = CamcopsColumn(
+        "q2_choice", Integer,
+        comment="Treatment preference (0 no, 1 yes)",
+        permitted_value_checker=ZERO_TO_ONE_CHECKER)
+    q3_choice = CamcopsColumn(
+        "q3_choice", Integer,
+        comment="Preference offered (0 no, 1 yes, 2 N/A)",
+        permitted_value_checker=ZERO_TO_TWO_CHECKER)
 
-    q1_satisfaction = CamcopsColumn("q1_satisfaction", Integer, comment="Patient satisfaction", permitted_value_checker=ZERO_TO_FOUR_CHECKER)
-    q2_satisfaction = CamcopsColumn("q2_satisfaction", UnicodeText, comment="Service expererience")
+    q1_satisfaction = CamcopsColumn(
+        "q1_satisfaction", Integer,
+        comment="Patient satisfaction",
+        permitted_value_checker=ZERO_TO_FOUR_CHECKER)
+    q2_satisfaction = CamcopsColumn(
+        "q2_satisfaction", UnicodeText,
+        comment="Service experience")
 
     MAIN_QUESTIONS = [
         "q_datetime",
         "q1_choice",
+        "q2_choice",
+        "q3_choice",
+        "q1_satisfaction",
     ]
 
     def is_complete(self) -> bool:
@@ -86,17 +106,11 @@ class Apeqpt(TaskHasPatientMixin, Task):
             return False
         return True
 
-    def get_trackers(self, req: CamcopsRequest) -> List[TrackerInfo]:
-        pass
-
-    def get_clinical_text(self, req: CamcopsRequest) -> List[CtvInfo]:
-        pass
-
     def get_summaries(self, req: CamcopsRequest) -> List[SummaryElement]:
         return self.standard_task_summary_fields()
 
     def get_task_html(self, req: CamcopsRequest) -> str:
-        c_dict= {
+        c_dict = {
             0: "0 — " + self.wxstring(req, "a0_choice"),
             1: "1 — " + self.wxstring(req, "a1_choice"),
             2: "2 — " + self.wxstring(req, "a2_choice"),
@@ -111,11 +125,14 @@ class Apeqpt(TaskHasPatientMixin, Task):
         q_a = ""
         for i in range(1, self.N_CHOICE_QUESTIONS + 1):
             nstr = str(i)
-            q_a += tr_qa(self.wxstring(req, "q" + nstr + "_choice"),
-                         get_from_dict(c_dict, getattr(self, "q" + nstr + "_choice")))
+            q_a += tr_qa(
+                self.wxstring(req, "q" + nstr + "_choice"),
+                get_from_dict(c_dict, getattr(self, "q" + nstr + "_choice")))
 
-        q_a += tr_qa(self.wxstring(req, "q1_satisfaction"), get_from_dict(s_dict, self.q1_satisfaction))
-        q_a += tr_qa(self.wxstring(req, "q2_satisfaction"), self.q2_satisfaction)
+        q_a += tr_qa(self.wxstring(req, "q1_satisfaction"),
+                     get_from_dict(s_dict, self.q1_satisfaction))
+        q_a += tr_qa(self.wxstring(req, "q2_satisfaction"),
+                     self.q2_satisfaction)
 
         h = """
             <div class="{CssClass.SUMMARY}">
@@ -124,8 +141,8 @@ class Apeqpt(TaskHasPatientMixin, Task):
                 </table>
             </div>
             <div class="{CssClass.EXPLANATION}">
-                Patient satisfaction rating rating for service provided. The service is rated on
-                choice offered and general satisfaction.
+                Patient satisfaction rating for service provided. The service
+                is rated on choice offered and general satisfaction.
             </div>
             <table class="{CssClass.TASKDETAIL}">
                 <tr>